+++ src/command/index.ts
@@ -2,6 +2,7 @@ enum CommandName {
     simpleTranslate = "simpleTranslate",
     completeTranslate = "completeTranslate",
     translation = "translation",
+    playSourceAudio = "playSourceAudio",
 }
 
 export {