+++ src/component/history.ts
@@ -4,8 +4,6 @@ import { App } from "../util/app"
 
 const key = "history"
 
-const historyMax = 50
-
 const readHistory = (): Array<api.TranslateItem> => {
     return (App.instance().getContext()?.globalState.get(key) ?? []) as Array<api.TranslateItem>
 }
@@ -16,7 +14,7 @@ const writeHistoryHelper = (content: Array<api.TranslateItem>) => {
 
 const writeHistory = (item: api.TranslateItem): Array<api.TranslateItem> => {
     let history = readHistory()
-    if (history.length >= historyMax) history.pop()
+    if (history.length >= (common.getUserConfig<number>(common.ConfigKey.maxHistory) ?? 20)) history.pop()
     history = history.filter(i => !(i.q === item.q && i.sl === item.sl && i.tl === item.tl))
     history.unshift(item)
     writeHistoryHelper(history)